[개발 #45] 영화 서버에 평균점수를 전달하는 API
평균 점수가 갱신되거나 삭제되거나 생길때 영화 서버에게 평균 점수를 전달한다
#45 참조

+++ chan/src/main/java/k5s/reviewdevelop/domain/Movie.java
@@ -21,9 +21,9 @@ public class Movie {
 
     private int num = 0;
 
-    private float sumScore = 0;
+    private double sumScore = 0;
 
-    private float averageScore;
+    private double averageScore;
 
     //==Movie가 삭제되면 review들도 삭제된다==//
     @OneToMany(mappedBy = "movie", cascade = CascadeType.ALL)
@@ -62,7 +62,7 @@ public class Movie {
             this.averageScore = (sumScore / num);
         }
         else {
-            this.averageScore = 0.00f; /**리뷰 갯수가 0일 경우 분모에 0이 올순 없다**/
+            this.averageScore = 0.00; /**리뷰 갯수가 0일 경우 분모에 0이 올순 없다**/
         }
     }
